move loader's queries to own module
