+++ packages/truffle-db/src/loaders/artifacts/index.ts
@@ -8,358 +8,14 @@ import Config from "@truffle/config";
 import { Environment } from "@truffle/environment";
 import Web3 from "web3";
 
-const AddBytecodes = gql`
-  input LinkReferenceInput {
-    offsets: [Int]
-    name: String
-    length: Int
-  }
-
-  input BytecodeInput {
-    bytes: Bytes!
-    linkReferences: [LinkReferenceInput]!
-  }
-
-  mutation AddBytecodes($bytecodes: [BytecodeInput!]!) {
-    workspace {
-      bytecodesAdd(input: { bytecodes: $bytecodes }) {
-        bytecodes {
-          id
-          linkReferences {
-            offsets
-            name
-            length
-          }
-        }
-      }
-    }
-  }
-`;
-
-const AddSources = gql`
-  input SourceInput {
-    contents: String!
-    sourcePath: String
-  }
-
-  mutation AddSource($sources: [SourceInput!]!) {
-    workspace {
-      sourcesAdd(input: { sources: $sources }) {
-        sources {
-          id
-          contents
-          sourcePath
-        }
-      }
-    }
-  }
-`;
-
-const AddCompilation = gql`
-  input CompilerInput {
-    name: String
-    version: String
-    settings: Object
-  }
-
-  input CompilationSourceInput {
-    id: ID!
-  }
-
-  input CompilationSourceContractSourceInput {
-    id: ID!
-  }
-
-  input CompilationSourceContractAstInput {
-    json: String!
-  }
-
-  input CompilationSourceMapInput {
-    json: String!
-  }
-
-  input CompilationSourceContractInput {
-    name: String
-    source: CompilationSourceContractSourceInput
-    ast: CompilationSourceContractAstInput
-  }
-
-  input CompilationInput {
-    compiler: CompilerInput!
-    contracts: [CompilationSourceContractInput!]
-    sources: [CompilationSourceInput!]!
-    sourceMaps: [CompilationSourceMapInput]
-  }
-  input CompilationsAddInput {
-    compilations: [CompilationInput!]!
-  }
-
-  mutation AddCompilation($compilations: [CompilationInput!]!) {
-    workspace {
-      compilationsAdd(input: { compilations: $compilations }) {
-        compilations {
-          id
-          compiler {
-            name
-            version
-          }
-          contracts {
-            name
-            source {
-              contents
-              sourcePath
-            }
-            ast {
-              json
-            }
-          }
-          sources {
-            contents
-            sourcePath
-          }
-          sourceMaps {
-            json
-          }
-        }
-      }
-    }
-  }
-`;
-
-const AddContracts = gql`
-  input AbiInput {
-    json: String!
-    items: [String]
-  }
-
-  input ContractCompilationInput {
-    id: ID!
-  }
-
-  input ContractSourceContractInput {
-    index: FileIndex
-  }
-
-  input ContractConstructorBytecodeInput {
-    id: ID!
-  }
-
-  input LinkReferenceInput {
-    bytecode: ID!
-    index: FileIndex
-  }
-
-  input ContractConstructorLinkValueInput {
-    value: Address!
-    LinkReference: LinkReferenceInput
-  }
-
-  input ContractConstructorLinkedBytecodeInput {
-    bytecode: ContractConstructorBytecodeInput!
-    linkValues: [ContractContructorLinkValueInput]
-  }
-
-  input ContractConstructorInput {
-    createBytecode: ContractConstructorLinkedBytecodeInput!
-  }
-
-  input ContractInput {
-    name: String
-    abi: AbiInput
-    compilation: ContractCompilationInput
-    sourceContract: ContractSourceContractInput
-    constructor: ContractConstructorInput
-  }
-
-  mutation AddContracts($contracts: [ContractInput!]!) {
-    workspace {
-      contractsAdd(input: { contracts: $contracts }) {
-        contracts {
-          id
-          name
-          abi {
-            json
-          }
-          sourceContract {
-            name
-            source {
-              contents
-              sourcePath
-            }
-            ast {
-              json
-            }
-          }
-          compilation {
-            compiler {
-              name
-              version
-            }
-            contracts {
-              name
-              source {
-                contents
-                sourcePath
-              }
-              ast {
-                json
-              }
-            }
-            sources {
-              contents
-              sourcePath
-            }
-          }
-          constructor {
-            createBytecode {
-              bytecode {
-                bytes
-                linkReferences {
-                  offsets
-                  name
-                  length
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-const AddContractInstances = gql`
-  input ContractInstanceNetworkInput {
-    id: ID!
-  }
-
-  input ContractInstanceContractInput {
-    id: ID!
-  }
-
-  input ContractInstanceCreationConstructorBytecodeInput {
-    id: ID!
-  }
-
-  input LinkReferenceInput {
-    offsets: [Int]
-    name: String
-    length: Int
-  }
-
-  input LinkValueLinkReferenceInput {
-    bytecode: ID!
-    index: FileIndex
-  }
-
-  input LinkValueInput {
-    value: Address!
-    linkReference: LinkValueLinkReferenceInput!
-  }
-
-  input ContractInstanceCreationConstructorLinkedBytecodeInput {
-    bytecode: ContractInstanceCreationConstructorBytecodeInput!
-    linkValues: [LinkValueInput]
-  }
-
-  input ContractInstanceCallBytecodeInput {
-    id: ID!
-  }
-
-  input ContractInstanceLinkedCallBytecodeInput {
-    bytecode: ContractInstanceBytecodeInput
-    linkValues: [LinkValueInput]
-  }
-
-  input ContractInstanceCreationConstructorInput {
-    createBytecode: ContractInstanceCreationConstructorLinkedBytecodeInput!
-  }
-
-  input ContractInstanceCreationInput {
-    transactionHash: TransactionHash!
-    constructor: ContractInstanceCreationConstructorInput!
-  }
-
-  input ContractInstanceInput {
-    address: Address!
-    network: ContractInstanceNetworkInput!
-    creation: ContractInstanceCreationInput
-    contract: ContractInstanceContractInput
-    callBytecode: ContractInstanceLinkedCallBytecodeInput
-  }
-
-  mutation AddContractInstances($contractInstances: [ContractInstanceInput!]!) {
-    workspace {
-      contractInstancesAdd(input: { contractInstances: $contractInstances }) {
-        contractInstances {
-          address
-          network {
-            name
-            networkID
-            historicBlock {
-              height
-              hash
-            }
-          }
-          contract {
-            name
-          }
-          creation {
-            transactionHash
-            constructor {
-              createBytecode {
-                bytecode {
-                  bytes
-                  linkReferences {
-                    offsets
-                    name
-                    length
-                  }
-                }
-                linkValues {
-                  value
-                  linkReference {
-                    offsets
-                    name
-                    length
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-const AddNetworks = gql`
-  input HistoricBlockInput {
-    height: Int!
-    hash: String!
-  }
-
-  input NetworkInput {
-    name: String
-    networkId: NetworkId!
-    historicBlock: HistoricBlockInput!
-  }
-
-  mutation AddNetworks($networks: [NetworkInput!]!) {
-    workspace {
-      networksAdd(input: { networks: $networks }) {
-        networks {
-          id
-          networkId
-          historicBlock {
-            height
-            hash
-          }
-        }
-      }
-    }
-  }
-`;
+import {
+  AddBytecodes,
+  AddSources,
+  AddCompilation,
+  AddContracts,
+  AddContractInstances,
+  AddNetworks
+} from "./queries";
 
 type WorkflowCompileResult = {
   compilations: {
