+++ packages/truffle-db/src/loaders/artifacts/queries/bytecodes.ts
@@ -1,0 +1,29 @@
+import gql from "graphql-tag";
+
+export const AddBytecodes = gql`
+  input LinkReferenceInput {
+    offsets: [Int]
+    name: String
+    length: Int
+  }
+
+  input BytecodeInput {
+    bytes: Bytes!
+    linkReferences: [LinkReferenceInput]!
+  }
+
+  mutation AddBytecodes($bytecodes: [BytecodeInput!]!) {
+    workspace {
+      bytecodesAdd(input: { bytecodes: $bytecodes }) {
+        bytecodes {
+          id
+          linkReferences {
+            offsets
+            name
+            length
+          }
+        }
+      }
+    }
+  }
+`;
