+++ packages/truffle-db/src/loaders/artifacts/queries/index.ts
@@ -1,0 +1,6 @@
+export { AddBytecodes } from "./bytecodes";
+export { AddSources } from "./sources";
+export { AddCompilation } from "./compilation";
+export { AddContracts } from "./contracts";
+export { AddContractInstances } from "./contractInstances";
+export { AddNetworks } from "./networks";
